created seperate function module to try to fix circular imports
